requirements.txt
Django>=5.0,<7.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
